Add instanceof check to prevent calling clearArg on non-block
For reasons not yet known, sometimes, when a block is dragged, it
drags an afterimage of the entire palette with it and moves when
the mouse is up. Then, to drop it, the mouse must be clicked. The
event hendler for mousedown assumes that the block is being
picked up, and an error occurs when the function tries to clear
the arg of the block's parent block. This commit fixes the
immediate problem, but I still need to prevent the dragging from
occurring incorrectly in the first place.